ref(backend): Migrate from Express to Hono + Bun; Closes #14 (#16)
