+++ .env
@@ -1,2 +1,2 @@
-VITE_BACKEND_API_URL = 'http://localhost:3000'
+VITE_BACKEND_API_URL='https://codesync-api.kshitizsharma.com.np'
 CORS_ORIGIN=*
