+++ codesync-server/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "codesync-server",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "dev": "bun --watch src/index.ts",
+    "start": "bun src/index.ts"
+  },
+  "dependencies": {
+    "@hono/node-server": "^1.13.7",
+    "hono": "^4.6.3",
+    "socket.io": "^4.8.1"
+  },
+  "devDependencies": {
+    "bun-types": "^1.1.0"
+  }
+}
